refactor(gui): decompose Electron main process
Split gui/src/main/index.ts (1,614 lines) into concern modules:
configStore, homeAssistant, aiSettings, voiceSettings, settingsDefaults,
settingsMirror, integrationStatus, integrationInventory, window, ipc,
plus handlers/settings, handlers/integrations, handlers/system following
the existing handlers/ register-function convention.

index.ts is now a 39-line entrypoint (app lifecycle wiring only). Code
moved verbatim; IPC channel names, payload shapes, bridge resolution,
and Electron security posture unchanged. Source-inspection tests updated
to read the moved modules; assertions unchanged.

US-REM-029.
