+++ docs/archive/progress/progress-remaining-release-readiness.txt
@@ -600,3 +600,11 @@ Use this format for each Ralph iteration:
 - The mypy ignore_errors exclusion block was already removed by US-REM-014 (commit 2a9facb); no core-module exclusions remain in pyproject.toml.
 - Verified after the US-REM-026/027/028 decompositions: mypy rex --ignore-missing-imports returns 0 errors (335 source files), including rex/cli.py, rex/voice_loop.py, rex/gui_app.py and the new rex/commands/ and rex/voice/ packages.
 - No code changes required; this entry records the validation evidence and closes the story checkboxes.
+
+## 2026-06-11 12:30 UTC - US-REM-029 complete
+- Decomposed gui/src/main/index.ts (1,614 lines) into gui/src/main/ concern modules: configStore (config/gui_settings/rex_config/.env file helpers), homeAssistant (HA credentials + HTTP client), aiSettings / voiceSettings / settingsDefaults (settings normalization, wake-word status, per-section defaults), settingsMirror (mirrorToRexConfig), integrationStatus (status store, fingerprints, configured-credential tests), integrationInventory (integration + capability inventory), window (createWindow), ipc (registerIpcHandlers aggregation).
+- Added handlers/settings.ts, handlers/integrations.ts, handlers/system.ts following the existing handlers/ register-function convention; all IPC channel names, argument shapes, and response payloads unchanged.
+- gui/src/main/index.ts is 39 lines: app.whenReady wiring only (validateBridges, settings mirror, createWindow, registerIpcHandlers, tray, window-all-closed). Electron security posture preserved: preload usage, window-open handler denies and opens externally, packaged-mode icon path, bridge resolution untouched.
+- Updated tests/test_outlook_integration_honesty.py and tests/test_us312_surface_settings.py to read the moved module sources (integrationStatus.ts, settingsMirror.ts); assertions unchanged. tests/test_us328_gui_launch_verification.py passes unmodified (index.ts retains validateBridges()/createWindow/app.whenReady/bridgeResolver).
+- Validation passed: npm run typecheck (both tsconfig projects), npm run build (electron-vite), electron-builder --dir repackage, SKIP_BUILD=1 bash tests/smoke/test_electron_package.sh (PASSED; packaged app logged "All bridge scripts validated successfully"), pytest on affected test files (47 passed), ruff check . clean, git diff --check clean.
+- Environment note: the smoke test initially failed because the local (gitignored) config/rex_config.json had models.llm_temperature stored as the string "0.7", making rex.config emit a stdout WARNING that broke the bridge JSON parse; fixed the local runtime value to a float per the warning text. Pre-existing condition unrelated to this story.
